Neural Network Example Output
This is an example of what a Neural Network output might look like.  The first column is seconds, the second column is whether an on/off trigger is detected [0/1], and the third is the the "meditation level". That meaning the last level is basically a score that goes up the longer meditation is detected.