Add files via upload
Some features are commented out for testing but you can select what recording you want to package, choose the point in the recording to start and stop recording (in seconds), and also the name for the output file. The input right now is designed around the output from the csv writer.